emotion_pp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
streamlit==1.28.1
python-multipart==0.0.6
aiofiles==23.2.1
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
kaggle==1.5.16
